+++ forge/test/models/pytorch/vision/regnet/utils/image_utils.py
@@ -7,6 +7,7 @@ from PIL import Image
 from transformers import AutoImageProcessor
 
 from test.utils import download_model
+import torch
 
 
 def preprocess_input_data(image_url, variant):
@@ -19,4 +20,4 @@ def preprocess_input_data(image_url, variant):
     # Preprocess the image
     image_tensor = preprocessor(images=image, return_tensors="pt").pixel_values
 
-    return [image_tensor]
+    return [image_tensor.to(torch.bfloat16)]
