+++ forge/test/models/pytorch/vision/yolo/utils/yolos_utils.py
@@ -9,7 +9,7 @@ from transformers import AutoImageProcessor, AutoModelForObjectDetection
 def load_model(variant):
     model = AutoModelForObjectDetection.from_pretrained(variant)
     model.eval()
-    return model
+    return model.to(torch.bfloat16)
 
 
 def load_input(variant):
@@ -17,4 +17,4 @@ def load_input(variant):
     image = Image.open(requests.get(test_input, stream=True).raw)
     image_processor = AutoImageProcessor.from_pretrained(variant)
     inputs = image_processor(images=image, return_tensors="pt")
-    return [inputs["pixel_values"]]
+    return [inputs["pixel_values"].to(torch.bfloat16)]
